--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LinearOpModeMotor.java
@@ -1,18 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-
-public class LinearOpModeMotor extends LinearOpMode {
-    private DcMotor motor;
-
-    @Override
-    public void runOpMode() {
-        motor = hardwareMap.get(DcMotor.class,"Motor");
-        waitForStart();
-    }
-
-}
-
-
-
