--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MotorTest/LinearOpModeMotor.java
@@ -1,20 +1,0 @@
-package org.firstinspires.ftc.teamcode.MotorTest;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-
-public class LinearOpModeMotor extends LinearOpMode {
-
-    private DcMotor motor;
-
-    @Override
-    public void runOpMode() {
-        motor = hardwareMap.get(DcMotor.class, "motor");
-
-        waitForStart();
-
-        while (opModeIsActive())
-            motor.setPower(0.5f);
-
-    }
-}
